+++ CHANGES.md
@@ -39,6 +39,7 @@
 * Remove defunct boxcar notifier
 * Add sound selection for boxcar2 notifier
 * Change boxcar2 notifier to use updated api scheme
+* Update the Plex notifier from a port at midgetspy/sickbeard
 
 [develop changelog]
 * Change uT params from unicode to str.format as magnet URLs worked but sending files in POST bodies failed